pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-railway-controller"
version = "0.1.0"
description = "Railway Traffic Controller Environment for OpenEnv - Train dispatching simulation for RL agents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "fastmcp>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
server = "railway_controller.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["railway_controller", "railway_controller.server"]
package-dir = { "railway_controller" = ".", "railway_controller.server" = "server" }